task: set "port" in parent implementation Task.get_template_args()
I think we should aim that the set of variables that we render is always
the same. That is, Task.get_template_args() renders most variables,
including "port".

Now, "port" should be set for ServerTask to the actual port number, and
for ClientTask to empty. This is achieved and there is no change in
behavior.

What we change here is that we also set the variable for "PluginTask"
subclasses (to empty). Sure, the PluginTask only renders
"tools-pod.yaml.j2", which does not use the variable, and, we anyway
only set it to empty.

Still, I find it clearer that we see what variables commonly exist
("port") and that we can easily find out who sets it to what
(_get_template_args_port()).

Signed-off-by: Thomas Haller <thaller@redhat.com>