task: let parent implementation Task.get_template_args() set "default_network"
The Task has a "node" attribute, which depending on the task's role is
either the node_server or the node_client.

We thus don't need subclasses to explicitly select which node. Just
always set the "default_network" variable.

Note that some subclasses of Task (PluginTask) probably don't have a use
for this field and don't need it when rendering their templates. In
practice, this is no problem to set a variable that is not used by the
template. Note how only "sriov-pod.yaml.j2' uses this variable, and this
template is only used by ServerTask/ClientTask (which previously also
set the Jinja2 variable in their get_template_args()). So, it should
work still the same.

Signed-off-by: Thomas Haller <thaller@redhat.com>