task: add "for_template" argument to cmd_line_args() methods
This will be useful later, when we unify things.

We have two cases where we construct the command line arguments, to put
into the pod manifest (with exec_persistent) or as thread action.

We will unify some code next and add an abstract method to the parent
class. In some cases there will be a difference which command line to
construct, so we will need this parameter later.

Signed-off-by: Thomas Haller <thaller@redhat.com>