all: use ktoolbox.common.run_main()
ktoolbox.common.run_main() runs the main function and logs a traceback
on any Exception. This is what we want for all our tools, because our
tools are not supposed to let exceptions escape, and if they do, we want
a traceback in the log. The helper does that. Use it.

Note that this changes the error code in some cases, as run_main() exits
with 70 (EX_SOFTWARE). That is an intentional change in behavior.

Signed-off-by: Thomas Haller <thaller@redhat.com>