task: use regular error codes instead of sys.exit(-1)
sys.exit(-1) really means os._exit(255), which looks a bit like exiting
due to signal 127 (no such signal exists). This is odd. Don't do this.
Instead, exit on error with a regular exit code (0-127).

Choose 70 (EX_SOFTWARE) and 78 (EX_CONFIG) from <sysexits.h> for this.

EX_SOFTWARE is also what ktoolbox.common.run_main() does and indicates a
bug-like situation (or something which we intentionally didn't handle
and are fine to abort as if it were a bug).

Signed-off-by: Thomas Haller <thaller@redhat.com>