task: fail ServerTask.confirm_server_alive() with exception
A plain sys.exit() in the middle of the code is not nice.
Also, this may or may not happen on a thread, where this may
or may not do the right thing.

Just raise an exception. We catch all exceptions at the highest level
and error out with a backtrace.

We either should make an effort to handle the error (e.g. continue the
test run and record an error in the test results), or we can just die
with an exception. Do the latter. But don't do it via sys.exit()
directly.

Also, `if not r or not r.success:` is redundant.

Signed-off-by: Thomas Haller <thaller@redhat.com>