tftbase: add str_sanitize() function for sanitizing
Preserves ASCII lower case letters and digits. But other characters
are escaped using "-".

The goal is that the input unambiguously maps to a result string (which
is reversible in theory).

Also, leading and trailing '-' are escaped too, because those are not
valid in a DNS label.

We will use this for the pod-name.

Signed-off-by: Thomas Haller <thaller@redhat.com>