testConfig: validate ConfNodeBase.name to be a valid DNS name
We use the node name to derive the pod_name. For that usage we have
"node_name_sanitize", which handles any characters.

Otherwise, we also use the node name for the pod selector:

  nodeSelector:
    kubernetes.io/hostname: {{ node_name }}

This hostname is usually the same as the node name. The node name must
be a valid DNS name. Technically, it seems you could configure the
hostname to differ from the node name, and then the hostname is not
required to be a valid DNS name. But let's not support that.

So this change breaks support for systems that rename the hostname of a
node to no longer be a valid DNS name. Just don't do that. Not
supported.

What this gives is that we know that the node name does not contain
bogus characters. For example, our naive Jinja2 replacement in the pod
files is not aware that it needs to insert a YAML string. Now we at
least know that the node name won't contain characters like '"' or '['
that would break the YAML.

Signed-off-by: Thomas Haller <thaller@redhat.com>