tftbase: fix str_sanitize() to handle zero unicode code point
We want that str_sanitize() maps any input string to a unique
replacement. Which means, the escaping is in principle reversable.

This was not the case for how "." and "\0000" was encoded. Fix that.
For example, the input strings "x\u0000Ax" and "x.0-ax" both were
escaped as "x-00--ax", a collision.

Fix that by using "-z" as unique prefix for escaping "Z" (becomes "-z-")
or hex strings (become now "-z<hex>-"). Basically, the escape sequences
must be prefix-free.

This changes behavior, but only for input with upper case "Z" or
non-regular characters. These are expected to not usually occur as input
where we use this function.

Signed-off-by: Thomas Haller <thaller@redhat.com>