Add a wait for listen socket before any traffic test client connects
We should make sure that the traffic test server is listening on
the tcp/udp socket first before allowing the client to connect. We do
this with all the different test types that are all vulnerable with this
type of race condition.

This makes sure that we don't get into any race conditions with a slow
start up of "iperf3 -s" or equilvalent.

Signed-off-by: William Zhao <wizhao@redhat.com>